+++ server_privileges.php
@@ -30,6 +30,7 @@ $scripts->addFile('server_privileges.js');
 if ((isset($_REQUEST['viewing_mode']) && $_REQUEST['viewing_mode'] == 'server')
     && $GLOBALS['cfgRelation']['menuswork']
 ) {
+    include_once 'libraries/server_users.lib.php';
     $response->addHTML('<div>');
     $response->addHTML(PMA_getHtmlForSubMenusOnUsersPage('server_privileges.php'));
 }
